Add entities script

+++ src/events/events.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Logger, Param, ParseIntPipe, Patch, Post, ValidationPipe } from "@nestjs/common";
+import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, ValidationPipe } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Like, MoreThan, Repository } from "typeorm";
 import { CreateEventDto } from "./create-event.dto";
@@ -23,24 +23,24 @@ export class EventsController {
         return events;
     }
 
-    // @Get('/practice') 
-    // async practice() {
-    //     return await this.repository.find({
-    //         select: ['id','when'],
-    //         where: 
-    //         [{
-    //             id: MoreThan(3),
-    //             when: MoreThan(new Date('2021-02-12T13:00:00'))
+    @Get('/practice') 
+    async practice() {
+        return await this.repository.find({
+            select: ['id','when'],
+            where: 
+            [{
+                id: MoreThan(3),
+                when: MoreThan(new Date('2021-02-12T13:00:00'))
 
-    //         },{
-    //             description: Like('%meet%')
-    //         }],
-    //         take:2 ,
-    //         order : {
-    //             id: 'DESC'
-    //         }
-    //     });
-    // }
+            },{
+                description: Like('%meet%')
+            }],
+            take:2 ,
+            order : {
+                id: 'DESC'
+            }
+        });
+    }
     
 
     @Get(':id')
@@ -48,7 +48,11 @@ export class EventsController {
         console.log(typeof id);
         // const event = this.events.find(event => event.id === parseInt(id));
         // return  event;
-        return await this.repository.findOne(id); 
+        const event =  await this.repository.findOne(id);
+        
+        if(!event) {
+            throw new NotFoundException();
+        }
         
     }  
 
@@ -86,6 +90,10 @@ export class EventsController {
         // return this.events[index];
 
         const event  = await this.repository.findOne(id);
+
+        if(!event) {
+            throw new NotFoundException();
+        }
         return await this.repository.save({
             ...event,
             ...input,
@@ -99,6 +107,9 @@ export class EventsController {
     async remove(@Param('id') id) {
         // this.events = this.events.filter(event => event.id !== parseInt(id));
         const event = await this.repository.findOne(id);
+        if(!event) {
+            throw new NotFoundException();
+        }
         await this.repository.remove(event);
     }
